requirements.txt
openai==0.28
tiktoken 
chromadb==0.4.14  #chromadb==0.4.14 
langchain==0.0.249
flask
GitPython
python-dotenv
-e .